requirements.txt
streamlit==1.44.1
pymongo==4.9.2
pandas==2.2.3
python-dotenv==1.0.1